안녕

+++ front-end/src/component/main/header.js
@@ -6,38 +6,56 @@ import '../../resource/css/login.css';
 
 function log(){
   const id = window.localStorage.getItem("id");
-  if(id == null){
-    console.log('로그인 안됨')
-    return <Nav >        
-        <Nav.Link href="/join" style={style}>회원가입</Nav.Link>
-        <Nav.Link eventKey={2} href="/login" style={style}>
-        로그인
+  const admin = window.localStorage.getItem("admin");
+  if(admin == null){
+    if(id == null){
+      console.log('로그인 안됨')
+      return <Nav >        
+          <Nav.Link href="/join" style={style}>회원가입</Nav.Link>
+          <Nav.Link eventKey={2} href="/login" style={style}>
+          로그인
+          </Nav.Link>
+      </Nav>
+    }
+    else if(id != null){
+      console.log('로그인 됨')
+      return <Nav>
+        <NavDropdown title={id} id="basic-nav-dropdown">
+          <NavDropdown.Item href="/memberInfo">내정보</NavDropdown.Item>
+          <NavDropdown.Item href="/modifyMember">회원수정</NavDropdown.Item>
+          <NavDropdown.Item href="/deleteMember">회원탈퇴</NavDropdown.Item>
+          <NavDropdown.Item href="/memAccount">회원 결산</NavDropdown.Item>
+          <NavDropdown.Item href="/question">1:1문의</NavDropdown.Item>
+        </NavDropdown>
+        <Nav.Link eventKey={2} href="/chat" style={style}>
+          채팅
         </Nav.Link>
-    </Nav>
+        <Nav.Link eventKey={2} style={style} onClick={logout}>
+          로그아웃
+        </Nav.Link>
+      </Nav>
+    }
   }
-  else if(id != null){
-    console.log('로그인 됨')
+  else if(admin != null){
     return <Nav>
-      <NavDropdown title={id} id="basic-nav-dropdown">
-      <NavDropdown.Item>--member--</NavDropdown.Item>
-        <NavDropdown.Item href="/memberInfo">내정보</NavDropdown.Item>
-        <NavDropdown.Item href="/modifyMember">회원수정</NavDropdown.Item>
-        <NavDropdown.Item href="/deleteMember">회원탈퇴</NavDropdown.Item>
-        <NavDropdown.Item href="/memAccount">회원 결산</NavDropdown.Item>
-        <NavDropdown.Item href="/question">1:1문의</NavDropdown.Item>
-      </NavDropdown>
-      <Nav.Link eventKey={2} href="/chat" style={style}>
-        채팅
-      </Nav.Link>
-      <Nav.Link eventKey={2} style={style} onClick={logout}>
-        로그아웃
-      </Nav.Link>
-    </Nav>
+        <NavDropdown title="관리자" id="basic-nav-dropdown">
+          <NavDropdown.Item href="/management">회원관리</NavDropdown.Item>
+          <NavDropdown.Item href="/adminAccount">관리자 결산</NavDropdown.Item>
+          <NavDropdown.Item href="/answer">1:1문의 답변</NavDropdown.Item>
+        </NavDropdown>
+        <Nav.Link eventKey={2} href="/chat" style={style}>
+          채팅
+        </Nav.Link>
+        <Nav.Link eventKey={2} style={style} onClick={logout}>
+          로그아웃
+        </Nav.Link>
+      </Nav>
   }
 }
 function logout(){
   window.location.reload();
   localStorage.removeItem("id");
+  localStorage.removeItem("admin");
 }
 
 function BasicExample() {
